analytics-service/requirements.txt
mysql-connector-python
pymongo
APScheduler
